refactor(store): shatter into small bits
