+++ src/store/getters.ts
@@ -1,0 +1,3 @@
+export default {
+  
+}
